Changed Windows symbol level to 1
On Windows, it is necessary to have symbols produced by the build because post-build
scripts appear to fail if there are no symbol files. As on Windows, symbols are not
part of the DLL but separate files, creating symbols does not increase DLL size and
thus is not a problem at all, so we can easily have them generated there.