Added Linux OpenJDK11 JVMs for native linking
This fixes a linking problem occurring with some JVMs on build servers which export JAWT_GetAWT in libjawt.so with a "SUNWprivate_1.1" version. Not all OpenJDK11+ VMs have this version added, which can cause runtime problems if compiling and linking against such a JVM with the version and having a JVM without it at runtime. The added JVM (Amazon Corretto) doesn't have the version and thus is safe to use for linking during compilation.